New jshint config file path and improved installation scripts.

+++ install.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-FILES=( .fonts .nvm .rbenv .oh-my-zsh .vim .vimrc .vimrc-global .zshrc )
+FILES=( .fonts .nvm .rbenv .oh-my-zsh .vim .vimrc .vimrc-global .zshrc .jshint )
 
 for file in ${FILES[@]}; do
   if [[ $file == ".vimrc-global" ]]; then
@@ -11,11 +11,21 @@ for file in ${FILES[@]}; do
 done
 
  ./python/install.sh
- ./nodejs-fetcher 0.12.0
 
- tar xvf *.tar.gz
- rm -rfv *.tar.gz
- cd *0.12.0
- ./configure && make && sudo make install
+if type node &>/dev/null; then
+  if type jshint &>/dev/null; then
+    exit 0
+  else
+    sudo npm install -g jshint
+  fi
+else
+  ./nodejs-fetcher 0.12.0
 
- sudo npm update -g
+  tar xvf *.tar.gz
+  rm -rfv *.tar.gz
+
+  cd *0.12.0
+  ./configure && make && sudo make install
+
+  sudo npm update -g && sudo npm install -g jshint
+fi
